% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{use_css}
\alias{use_css}
\title{Add CSS}
\usage{
use_css()
}
\description{
This function adds the shared CSS files by creating a resource path to
the package and linking to the CSS stylesheet in the HTML head.
}
