% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_functions.R
\name{use_js}
\alias{use_js}
\title{Use javascript functions}
\usage{
use_js()
}
\value{
A \code{script} tag to be placed in the \code{ui} function, \code{ui.R} file.
}
\description{
Vicrentaldash package includes a few javascript functions to supplement the
available shiny functions. This function will import and declare them.
}
