% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vicrentaldash-package.R
\docType{package}
\name{vicrentaldash-package}
\alias{vicrentaldash}
\alias{vicrentaldash-package}
\title{vicrentaldash: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Miguel Ayala \email{miguel.ayala.au@gmail.com}

}
\keyword{internal}
